postgres implementation

+++ project/apps/feedback/src/app/feedback.service.ts
@@ -6,7 +6,7 @@ import {
 import { CreateCommentDto } from './dtos';
 import { FeedbackRepository } from './feedback.repository';
 import { CommentEntity } from './comment.entity';
-import { ErrorMessages } from '@project/core';
+import { ErrorMessages, PostState } from '@project/core';
 import { PostEntity } from '../../../posts/src/app/post.entity';
 import { LikeEntity } from './like.entity';
 
@@ -14,7 +14,7 @@ import { LikeEntity } from './like.entity';
 export class FeedbackService {
   constructor(private feedbackRepository: FeedbackRepository) {}
 
-  public async create(
+  public async createComment(
     userId: string,
     dto: CreateCommentDto
   ): Promise<CommentEntity> {
@@ -27,7 +27,7 @@ export class FeedbackService {
     return this.feedbackRepository.findCommentsByPostId(postId);
   }
 
-  public async delete(userId: string, id: string) {
+  public async deleteComment(userId: string, id: string) {
     const comment = await this.feedbackRepository.findCommentById(id);
     if (comment) {
       if (comment.createdBy === userId)
@@ -38,12 +38,25 @@ export class FeedbackService {
   }
 
   public async like(userId: string, postId: string): Promise<void> {
-    const postEntity = await this.feedbackRepository.findLike(userId, postId);
+    const postEntity = await this.feedbackRepository.client.post.findUnique({
+      where: { id: postId },
+    });
+
     if (postEntity) {
-      return await this.feedbackRepository.deleteLike(userId, postId);
+      if (postEntity.state === PostState.Published) {
+        const likeEntity = await this.feedbackRepository.findLike(
+          userId,
+          postId
+        );
+        if (likeEntity) {
+          return await this.feedbackRepository.deleteLike(userId, postId);
+        }
+        return await this.feedbackRepository.saveLike(
+          new LikeEntity({ postId, createdBy: userId })
+        );
+      }
+      throw new BadRequestException(ErrorMessages.PostNotPublish);
     }
-    return await this.feedbackRepository.saveLike(
-      new LikeEntity({ postId, createdBy: userId })
-    );
+    throw new NotFoundException(ErrorMessages.PostNotFound);
   }
 }
